requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
anyio==3.4.0; python_full_version >= '3.6.2'
apscheduler==3.8.1
asgiref==3.4.1; python_version >= '3.6'
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.9; python_version >= '3'
click==8.0.3; python_version >= '3.6'
dnspython==2.1.0; python_version >= '3.6'
email-validator==1.1.3
fastapi[all]==0.70.1
graphql-core==3.1.7; python_version >= '3.6' and python_version < '4.0'
h11==0.12.0; python_version >= '3.6'
httptools==0.2.0
idna==3.3; python_version >= '3'
itsdangerous==2.0.1
jinja2==3.0.3
markupsafe==2.0.1; python_version >= '3.6'
motor==2.5.1
orjson==3.6.5
pydantic==1.8.2; python_full_version >= '3.6.1'
pygments==2.10.0; python_version >= '3.5'
pymongo==3.12.3
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.19.2
python-multipart==0.0.5
pytz-deprecation-shim==0.1.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pytz==2021.3
pyyaml==5.4.1
requests==2.26.0
sentinel==0.3.0; python_version >= '3.6' and python_version < '4.0'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
starlette==0.16.0; python_version >= '3.6'
strawberry-graphql[fastapi]==0.93.5
typing-extensions==4.0.1; python_version >= '3.6'
tzdata==2021.5; python_version >= '3.6'
tzlocal==4.1; python_version >= '3.6'
ujson==4.3.0
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4.0'
uvicorn[standard]==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
xmltodict==0.12.0
